+++ CHANGELOG.md
@@ -57,6 +57,7 @@ To use this you'll need compatible versions:
 -   Added TOTP recipe. For more details please check our guide [here](TODO)
 -   Fixed a font loading issue, that caused apps using the default (Rubik) font to appear with the incorrect font weights
 -   Some default styling has changed related to how fonts/font-weights are applied
+-   Changed how `headerSubtitle` is styled in components: `EmailPasswordResetPasswordEmail`, `EmailPasswordSubmitNewPassword`, `EmailPasswordSignInHeader`, `EmailPasswordSignUpHeader`
 
 ## [0.35.9] - 2023-12-06